% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barVote.R
\name{barVote}
\alias{barVote}
\title{barVote}
\usage{
barVote(tipo, data, vx, vy, grupo, titulo, reg = FALSE)
}
\arguments{
\item{tipo}{{The type of votes: General, Sin, Superior}}

\item{data}{{The data to be displayed}}

\item{vx}{{Variable in the x-axis}}

\item{vy}{{Variable in the y-axis}}

\item{grupo}{{The grouping variable}}

\item{titulo}{{The title of the table}}

\item{reg}{{Specifying region}}
}
\value{

}
\description{
A barplot to visualize voting intention
}
\details{
barVote
}
