% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataClean.R
\name{dataClean}
\alias{dataClean}
\title{Consulta Saldo: dataClean}
\usage{
dataClean(
data = survey,
email_column = 4,
)
}
\arguments{
\item{data}{{Data from survey}}

\item{email_column}{{Variable about participant email}}
}
\value{

}
\description{
Delete duplicates in user and cellphone columns
}
\details{
dataClean
}
