% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateSurvey.R
\name{dateSurvey}
\alias{dateSurvey}
\title{dateSurvey}
\usage{
dateSurvey(data, fecha)
}
\arguments{
\item{data}{{The data to be displayed}}

\item{fecha}{{Date information about survey}}
}
\value{

}
\description{
A function to create surveys' date
}
\details{
dateSurvey
}
\examples{

}
