% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descMo.R
\name{descMo}
\alias{descMo}
\title{IDEA: descMo}
\usage{
descMo(mociones)
}
\arguments{
\item{proyectos}{{Database with motions by Congress of the Republic of Peru}}
}
\value{

}
\description{
Descriptive table about the motions by Congress of the Republic of Peru. Motions are divided into principal author, type, last state and parliamentary group
}
\details{
descMo
}
\examples{

}
