% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descPy.R
\name{descPy}
\alias{descPy}
\title{IDEA: descPy}
\usage{
descPy(proyectos)
}
\arguments{
\item{proyectos}{{Database with law projects by Congress of the Republic of Peru}}
}
\value{

}
\description{
Descriptive table about the projects by Congress of the Republic of Peru. Projects are divided into principal author, last state and parliamentary group
}
\details{
descPy
}
\examples{

}
