% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exjson.R
\name{exjson}
\alias{exjson}
\title{Consulta Saldo: exjson}
\usage{
exjson(data, name)
}
\arguments{
\item{data}{{Output of the "dataFinal" function}}

\item{name}{{Name of exported file}}
}
\value{

}
\description{
Converts survey's data to a json format and saves it with the same name
}
\details{
exjson
}
\examples{

}
