% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqAttribute.R
\name{freqAttribute}
\alias{freqAttribute}
\title{Consulta Saldo: freqAttribute}
\usage{
freqAttribute(data, columns, rango)
}
\arguments{
\item{data}{{Survey data}}

\item{columns}{{Position of attributes}}

\item{rango}{{Numbers of attributes}}
}
\value{

}
\description{
Proportion table by attribute's order
}
\details{
freqAttribute
}
\examples{

}
