% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqLevel.R
\name{freqLevel}
\alias{freqLevel}
\title{Consulta Saldo: freqLevel}
\usage{
freqLevel(data, rango)
}
\arguments{
\item{data}{{Survey data}}

\item{rango}{{Numbers of levels}}
}
\value{

}
\description{
Proportion table by levels
}
\details{
freqLevel
}
