% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUsers.R
\name{getUsers}
\alias{getUsers}
\title{Consulta Saldo: getUsers}
\usage{
getUsers(dataClean)
}
\arguments{
\item{dataClean}{{"dataClean" function's output}}
}
\value{

}
\description{
Get user variable by survey participants. It's linked with "dataClean" function
}
\details{
getUsers
}
\examples{

}
