% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUsersCleaned.R
\name{getUsersCleaned}
\alias{getUsersCleaned}
\title{Consulta Saldo: getUsersCleaned}
\usage{
getUsersCleaned(dataLimpia)
}
\arguments{
\item{dataClean}{{"dataClean" function's output}}
}
\value{

}
\description{
Get user variable by survey participants and delete duplicates. It's linked with "dataClean" function
}
\details{
getUsersCleaned
}
\examples{

}
