% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcion2.R
\name{linea}
\alias{linea}
\title{Gráfico temporal}
\usage{
linea(data, categoria, var, grupo, titulo)
}
\arguments{
\item{data}{}
\item{categoria}{}
\item{var}{}
\item{grupo}{}
\item{titulo}{}
}
\value{

}
\description{
Gráfico temporal
}
