% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longtable.R
\name{longtable.stargazer}
\alias{longtable.stargazer}
\title{longtable}
\usage{
plotLineVoting(
...             = ...,
float           = TRUE,
longtable.float = FALSE,
longtable.head  = TRUE,
filename        = NULL
)
}
\arguments{
\item{...}{{analysis}}

\item{float}{{floating environment}}

\item{longtable.float}{{floating environment}}

\item{longtable.head}{{head of final table}}

\item{filename}{{exporting name}}
}
\value{

}
\description{
A table to integrate several analysis's output
}
\details{
longtable
}
\examples{

}
