% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nineDigits.R
\name{nineDigits}
\alias{nineDigits}
\title{Consulta Saldo: nineDigits}
\usage{
nineDigits(
x = data$cellphone,
n = 9,
)
}
\arguments{
\item{x}{{cell phone variable}}

\item{n}{{digits to select}}
}
\value{

}
\description{
Select the nine last digits from cell phone variable
}
\details{
nineDigits
}
\examples{

}
