% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_matrix.R
\name{percent_matrix}
\alias{percent_matrix}
\title{percent_matrix}
\usage{
percent_matrix(
data  = data,
v1    = data$variable1,
v2    = data$variable2,
break_limit = 5
)
}
\arguments{
\item{data}{{The data to be displayed}}

\item{v1}{{The first parameter in matrix}}

\item{v2}{{The second parameter in matrix}}

\item{break_limit}{{number of characters to split the text}}
}
\value{

}
\description{
Calculate percentages by questions with matrix style
}
\details{
percent_matrix
}
