% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_multiple.R
\name{percent_multiple}
\alias{percent_multiple}
\title{percent_multiple}
\usage{
percent_multiple(
data      = data,
variable  = data$variable
)
}
\arguments{
\item{data}{{The data to be displayed}}

\item{variable}{{The variable to analyze}}
}
\value{

}
\description{
Calculate percentages by multiple answer questions
}
\details{
percent_multiple
}
