% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_simple.R
\name{percent_simple}
\alias{percent_simple}
\title{percent_simple: a table percentage}
\usage{
percent_simple(
data      = data,
variable  = data$variable
vgrupo    = NULL
vreg      = NULL
)
}
\arguments{
\item{data}{{The data to be displayed}}

\item{variable}{{The variable to analyze}}

\item{vgrupo}{{The grouping variable}}

\item{vreg}{{Specifying region}}
}
\value{

}
\description{
Calculate percentages by questions with single answer
}
\details{
percent_simple
}
