% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBar.R
\name{plotBar}
\alias{plotBar}
\title{plotBar}
\usage{
plotBar(
data      = survey,
vx        = survey$variable1,
vy        = survey$Porc,
titulo    = "Title of the graphic",
grupo     = NULL,
filp      = FALSE
)
}
\arguments{
\item{data}{{The data to be displayed in this layer}}

\item{vx}{{Variable in x axis. Type variable: character}}

\item{vy}{{Variable in y axis. Contains the proportions from vx}}

\item{titulo}{{The title of the layer}}

\item{grupo}{{The grouping variable}}

\item{flip}{{The orientation of the layer}}
}
\value{

}
\description{
A graphic to visualize frequency of answers. In adittion, sorts elements by frecuency
}
\details{
plotBar
}
\examples{

}
