% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarStacked.R
\name{plotBarStacked}
\alias{plotBarStacked}
\title{plotBarStacked}
\usage{
plotBarStacked(
data      = survey,
vx        = survey$variable1,
vy        = survey$variable2,
titulo    = "Title of the graphic",
grupo     = survey$group,
filp      = FALSE
)
}
\arguments{
\item{data}{{The data to be displayed in this layer}}

\item{vx}{{Variable in x axis. Type variable: character}}

\item{vy}{{Variable in y axis. Contains the proportions from vx}}

\item{titulo}{{The title of the layer}}

\item{grupo}{{The grouping variable}}

\item{flip}{{The orientation of the layer}}
}
\value{

}
\description{
A graphic to visualize frequency of answers in a stacked format
}
\details{
plotBarStacked
}
\examples{

}
