% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarVoting.R
\name{plotBarVoting}
\alias{plotBarVoting}
\title{plotBarVoting}
\usage{
plotBarVoting(
data      = survey,
vx        = survey$variable1,
vy        = survey$variable2,
titulo    = "Title of the graphic",
)
}
\arguments{
\item{data}{{The voting data to be displayed in this layer}}

\item{vx}{{Variable in x axis}}

\item{vy}{{Variable in y axis}}

\item{titulo}{{The title of the layer}}
}
\value{

}
\description{
A graphic to visualize voting intention. Shows a barplot with names by candidates and political parties
}
\details{
plotBarVoting
}
\examples{

}
