% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFreq.R
\name{plotFreq}
\alias{plotFreq}
\title{IDEA: plotFreq}
\usage{
plotFreq(
data      = database,
variable  = "autor",
frec_min  = 1,
titulo    = "Title of the graphic"
)
}
\arguments{
\item{data}{{Database to be displayed}}

\item{variable}{{A charactheristic about project or motion}}

\item{frec_min}{{Minimum frequency number to show in the plot}}

\item{titulo}{{Title of the plot}}
}
\value{

}
\description{
Barplot about projects and motions by Congress of the Republic of Peru
}
\details{
plotFreq
}
\examples{

}
