% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotline.R
\name{plotline}
\alias{plotline}
\title{plotline: A graphic to analyze time evolution}
\usage{
plotline(
data = survey,
categoria = "general",
var = "ads",
grupo = "proyecto",
titulo = "Title of the graphic"
)
}
\arguments{
\item{data}{{The data to be displayed in this layer}}

\item{categoria}{{Ads category.There are two options:
If "general", data is conformed by only ads to general audience.
If "hombres", data is conformed by only ads to male audience.}}

\item{var}{{The variable to analyze}}

\item{grupo}{{The grouping variable}}

\item{titulo}{{The title of the layer.}}
}
\value{

}
\description{
A graphic to analyze time evolution of some survey variable
}
\details{
plotline
}
