% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLineVoting.R
\name{plotLineVoting}
\alias{plotLineVoting}
\title{plotlineVoting}
\usage{
plotLineVoting(data, vx, vy, grupo, dates, titulo)
}
\arguments{
\item{data}{{The data to be displayed in this layer}}

\item{vy}{{Variable in the y-axis. Contains information about responses}}

\item{grupo}{{The grouping variable}}

\item{dates}{{Date information}}

\item{titulo}{{The title of the layer}}
}
\value{

}
\description{
A graphic to analyze time evolution of approval ratings
}
\details{
plotlineVoting
}
\examples{

}
