% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{processData}
\alias{processData}
\title{Consulta Saldo: processData}
\usage{
processData(data, ganador, fecha)
}
\arguments{
\item{data}{{Data survey}}

\item{ganador}{{User of survey winner}}

\item{fecha}{{Date information about survey}}
}
\value{

}
\description{
Shows a table with absolute frequencies by users and surveys
}
\details{
processData
}
\examples{

}
