% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFinal.R
\name{processFinal}
\alias{processFinal}
\title{Consulta Saldo: processFinal}
\usage{
processFinal(data, ganador, fecha, encuesta)
}
\arguments{
\item{data}{{Data survey}}

\item{ganador}{{User of survey winner}}

\item{fecha}{{Date information about survey}}

\item{encuesta}{{Recent survey}}
}
\value{

}
\description{
Joins processed tables of previous polls with the recent one
}
\details{
processFinal
}
\examples{

}
