% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_accent.R
\name{rm_accent}
\alias{rm_accent}
\title{IDEA: rm_accent}
\usage{
rm_accent(var = data$variable)
}
\arguments{
\item{var}{{Variable. Type: character or factor}}
}
\value{

}
\description{
Removes all accents from characters in selected variable
}
\details{
rm_accent
}
\examples{

}
