% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardVar.R
\name{standardVar}
\alias{standardVar}
\title{IDEA: standardVar}
\usage{
standardVar(
data = data,
var  = data$variable,
n    = NULL,
)
}
\arguments{
\item{data}{{Database}}

\item{var}{{Variable}}

\item{n}{{Validates type of the "var" parameter. n == 1 if "var" is string and n > 1 if "var" is list}}
}
\value{

}
\description{
Removes all unusual characters (quotation marks & capital letters) in selected variable
}
\details{
standardVar
}
\examples{

}
