% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsNotFin.R
\name{statsNotFin}
\alias{statsNotFin}
\title{Consulta Saldo: statsNotFin}
\usage{
statsNotFin(name_data, excluir_names, excluir_range, n_columns)
}
\arguments{
\item{name_data}{{Data survey}}

\item{excluir_names}{{Exclude conjoint sections}}

\item{excluir_range}{{Joins variables with NA's }}

\item{n_columns}{{Number of columns}}
}
\value{

}
\description{
Customizes conjoint data. Function shows the specific number of columns
}
\details{
statsNotFin
}
