% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsTime.R
\name{statsTime}
\alias{statsTime}
\title{Consulta Saldo: statsTime}
\usage{
statsTime(data)
}
\arguments{
\item{data}{{Output of the "timeSection" function}}
}
\value{

}
\description{
Shows a table with descriptive statistics by  sections of the poll
}
\details{
statsTime
}
\examples{

}
