% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabFinal.R
\name{tabFinal}
\alias{tabFinal}
\title{Consulta Saldo: tabFinal}
\usage{
tabFinal(data, retiros)
}
\arguments{
\item{data}{{processed data}}

\item{retiros}{{withdrawals colum}}
}
\value{

}
\description{
Groups processed data by users, and joins it with withdrawals' table
}
\details{
tabFinal
}
