% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSection.R
\name{timeSection}
\alias{timeSection}
\title{Consulta Saldo: timeSection}
\usage{
timeSection(data, var1, var2, submit_bloque1)
}
\arguments{
\item{data}{{Survey data}}

\item{var1}{{Question's label}}

\item{var2}{{Frecuency of responses by question}}

\item{submit_bloque1}{{Column's position}}
}
\value{

}
\description{
Calculates completion time by users in every page break of qualtrics
}
\details{
timeSection
}
\examples{

}
