% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalVaccine.R
\name{totalVaccine}
\alias{totalVaccine}
\title{totalVaccine}
\usage{
totalVaccine()
}
\value{

}
\description{
A function to merge the vaccine surveys and create a summary by survey's date
}
\details{
totalVaccine
}
\examples{

}
