% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimmer.R
\name{trimmer}
\alias{trimmer}
\title{trimmer}
\usage{
trimmer(
x            = data$column,
break_limit  = 10
)
}
\arguments{
\item{x}{{column or variable}}

\item{break_limit}{{number of characters to split the text}}
}
\value{

}
\description{
Adds '\n' after an group of specified characthers
}
\details{
trimmer
}
