% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgradePyMo.R
\name{upgradePyMo}
\alias{upgradePyMo}
\title{IDEA: upgradePyMo}
\usage{
upgradePyMo(proyectos, mociones, upgrade)
}
\arguments{
\item{proyectos}{{Database with law projects by Congress of the Republic of Peru}}

\item{mociones}{{Database with political motions by Congress of the Republic of Peru}}

\item{upgrade}{{Database to be upgrade, called "TablaClasificarFinal"}}
}
\value{

}
\description{
Joins new projects and motions in "TablaClasificarFinal.xlsx" and shows the mark of them (new, registered, eliminated)
}
\details{
upgradePyMo
}
\examples{

}
