% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote.R
\name{vote}
\alias{vote}
\title{vote}
\usage{
vote(tipo, data, vx, vy, grupo, dates, titulo, reg = FALSE)
}
\arguments{
\item{tipo}{{The type of votes: General, Sin, Superior}}

\item{data}{{The data to be displayed}}

\item{vx}{{Variable in the x-axis}}

\item{vy}{{Variable in the y-axis}}

\item{grupo}{{The grouping variable}}

\item{dates}{{Date information}}

\item{titulo}{{The title of the table}}

\item{reg}{{Specifying region}}
}
\value{

}
\description{
Calculate percentages by votes
}
\details{
vote
}
