% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcion2.R, R/funcion3.R, R/plotline.R
\name{linea}
\alias{linea}
\title{Gráfico temporal}
\usage{
linea(data, categoria, var, grupo, titulo)

linea(data, categoria, var, grupo, titulo)

linea(data, categoria, var, grupo, titulo)
}
\arguments{
\item{titulo}{}
}
\value{





}
\description{
Gráfico temporal

Title

Title
}
\examples{

}
