% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBar.R
\name{plotBar}
\alias{plotBar}
\title{plotBar}
\usage{
plotBar(data, vx, vy = Porc, titulo, grupo = NULL, flip = FALSE)
}
\arguments{
\item{data}{{The data to be displayed in this layer}}

\item{vx}{{Variable in x axis}}

\item{vy}{{Variable in y axis}}

\item{titulo}{{The title of the layer}}

\item{grupo}{{The grouping variable}}

\item{flip}{{The orientation of the layer}}
}
\value{

}
\description{
A graphic to visualize frequency of answers
}
\details{
plotBar
}
\examples{

}
