% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarUnsort.R
\name{plotBarUnsort}
\alias{plotBarUnsort}
\title{plotBarUnsort}
\usage{
plotBar(
data      = survey,
vx        = survey$variable1,
vy        = survey$variable2,
titulo    = "Title of the graphic",
grupo     = NULL,
filp      = FALSE
)
}
\arguments{
\item{data}{{The data to be displayed in this layer}}

\item{vx}{{Variable in x axis}}

\item{vy}{{Variable in y axis}}

\item{titulo}{{The title of the layer}}

\item{grupo}{{The grouping variable}}

\item{flip}{{The orientation of the layer}}
}
\value{

}
\description{
A graphic to visualize frequency of answers, but it does not sort elements by frecuency
}
\details{
plotBarUnsort
}
