% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLineVoting.R
\name{plotLineVoting}
\alias{plotLineVoting}
\title{plotLineVoting}
\usage{
plotLineVoting(
data      = survey,
vx        = survey$variable1,
vy        = survey$variable2,
grupo     = survey$group,
dates     = survey$date,
title     = "Title of the graphic"
)
}
\arguments{
\item{data}{{The data to be displayed in this layer}}

\item{vx}{{Variable in the x-axis. Contains information about survey}}

\item{vy}{{Variable in the y-axis. Contains information about responses}}

\item{grupo}{{The grouping variable}}

\item{dates}{{Date information}}

\item{titulo}{{The title of the layer}}
}
\value{

}
\description{
A graphic to analyze time evolution of approval ratings
}
\details{
plotLineVoting
}
\examples{

}
